test/pylib: move async query wrapper to common module
Move async query wrapper code out of topology and into its own module.
